backend/functions/gmail_watch/requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.*
google-auth==2.*
google-api-python-client==2.*
google-cloud-storage==2.*
